+++ app/components/base.rb
@@ -8,6 +8,7 @@ module Components
     include Phlex::Rails::Helpers::LinkTo
     include Phlex::Rails::Helpers::T
     include Phlex::Rails::Helpers::FormWith
+    include Phlex::Rails::Helpers::Pluralize
     include Components::FormHelpers
 
     if Rails.env.development?